Update and rename strip_ech addon to use new DNS HTTPS records API
